Refactoring Euphonic MVC.
unification of plot and settings widgets, as well as polymorphism instead of inheritance
and composition. This will help maintaining the code clear and readable.

This is still a draft commit, let's say. I started doing the unification in the euphonicmodel.py
and in the structurefactorwidget.py - but the files, imports and everything need to be fixed